requirements.txt
psycopg2-binary>=2.9.5
dashscope>=2.0.0
numpy>=1.24.0
pgvector>=0.2.0
pytest>=7.0.0